+++ src/useField.ts
@@ -3,17 +3,33 @@ import { useCallback, useEffect, useState, ChangeEvent } from 'react';
 import { get } from './utils';
 import useFormContext from './useFormContext';
 
+type State = {
+  value: any;
+  error: string | null;
+  touched: boolean;
+};
+
+const equalState = (a: State, b: State) =>
+  a.value === b.value && a.error === b.error && a.touched === b.touched;
+
 const useField = (name: string) => {
   const form = useFormContext();
-  const [value, setValue] = useState(get(form.values, name));
-  const [error, setError] = useState(form.errors[name] || null);
-  const [touched, setTouched] = useState(form.touched[name] || false);
+
+  const getState = () => ({
+    value: get(form.values, name),
+    error: form.errors[name] || null,
+    touched: form.touched[name] || false,
+  });
+
+  const [state, setState] = useState(getState);
 
   useEffect(() => {
     const unsubscribe = form.subscribe(() => {
-      setValue(get(form.values, name));
-      setError(form.errors[name] || null);
-      setTouched(form.touched[name] || false);
+      const newState = getState();
+
+      setState((prevState) =>
+        equalState(prevState, newState) ? prevState : newState,
+      );
     });
 
     return unsubscribe;
@@ -52,8 +68,8 @@ const useField = (name: string) => {
   }, []);
 
   return {
-    field: { name, value, onChange, onBlur },
-    meta: { error, touched },
+    field: { name, value: state.value, onChange, onBlur },
+    meta: { error: state.error, touched: state.touched },
     helpers: {
       setValue: setValueHelper,
       setTouched: setTouchedHelper,